Phase 1: Fix critical paper discrepancies
1. SIDRAViewConstraint: uint8_t → int32_t for window/ttl/refresh/min_agg
   (uint8_t max 255 caused overflow for values > 255, DB schema already uses INT)

2. Throw ParserException if client_id not found when min_agg > 0
   (was silently skipping min_agg enforcement, violating Eq. 12 privacy guarantee)

3. Validate single annotation per column (SENSITIVE, FACT, DIMENSION mutually
   exclusive per paper Section 3.3 privacy levels)

Tests: 219 assertions including new tests for multi-annotation rejection
and large WINDOW/TTL/REFRESH values.

+++ src/include/helpers.hpp
@@ -13,10 +13,10 @@ struct SIDRAConstraints {
 };
 
 struct SIDRAViewConstraint {
-	uint8_t window = 0;
-	uint8_t ttl = 0;
-	uint8_t refresh = 0;
-	uint8_t min_agg = 0;
+	int32_t window = 0;
+	int32_t ttl = 0;
+	int32_t refresh = 0;
+	int32_t min_agg = 0;
 };
 
 struct SIDRASelectOption {
